remove ddl worker
